using bfloat16 + updated SmolVLM2 rust-py interop

+++ vlmeval/vlm_custom/smolvlm2_py.py
@@ -28,7 +28,7 @@ class SmolVLM2PyComparison(BaseModel):
         self.processor = AutoProcessor.from_pretrained(model_path)
         self.model = AutoModelForImageTextToText.from_pretrained(
             model_path,
-            torch_dtype=torch.float32,
+            torch_dtype=torch.bfloat16,
             _attn_implementation="eager",
         ).to("cuda")
